backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiofiles>=24.1.0
pillow>=11.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
pytest-asyncio>=0.24.0
